pkgdown/mathjax-config.html

tulpaMesh: Constrained Delaunay Triangulation Meshes for Spatial 'SPDE' Models

Generate constrained Delaunay triangulation meshes for use with stochastic partial differential equation (SPDE) spatial models ('Lindgren', 'Rue' and 'Lindstroem' 2011 doi:10.1111/j.1467-9868.2011.00777.x ). Provides automatic mesh generation from point coordinates with boundary constraints, 'Ruppert' refinement for mesh quality, finite element method (FEM) matrix assembly (mass, stiffness, projection), barrier models, spherical meshes via icosahedral subdivision, and metric graph meshes for network geometries. Built on the 'CDT' header-only 'C++' library ('Amirkhanov' 2024 https://github.com/artem-ogre/CDT). Designed as the mesh backend for the 'tulpa' Bayesian hierarchical modelling engine but usable standalone for any spatial triangulation task.

See also

Author

Maintainer: Gilles Colling gilles.colling051@gmail.com (ORCID) [copyright holder]